+++ subpackages/cli/source/odood/cli/app.d
@@ -12,7 +12,6 @@ private import odood.exception: OdoodException;
 private import odood.cli.core.logger: OdoodLogger;
 private import odood.cli.core: OdoodProgram, OdoodCommand;
 private import odood.cli.commands.init: CommandInit;
-private import odood.cli.commands.deploy: CommandDeploy;
 private import odood.cli.commands.server:
     CommandServer, CommandServerStart, CommandServerStop, CommandServerRestart,
     CommandServerBrowse, CommandServerLogView;
@@ -30,6 +29,9 @@ private import odood.cli.commands.psql: CommandPSQL;
 private import odood.cli.commands.info: CommandInfo;
 private import odood.cli.commands.odoo: CommandOdoo;
 private import odood.cli.commands.precommit: CommandPreCommit;
+
+// Deploy is available only on Linux
+version(linux) private import odood.cli.commands.deploy: CommandDeploy;
 
 version(OdoodInDocker) {
     private import odood.lib.project: Project;
@@ -66,6 +68,7 @@ version(OdoodInDocker) {
             }
 
             if (args.argsRest.empty)
+                // TODO: May be better would be to call some func or method, instead of running self executable again?
                 return Process(thisExePath).withArgs("server", "run").execv;
             else if (args.argsRest[0] == "odood")
                 return Process(thisExePath).withArgs(args.argsRest[1 .. $]).execv;
@@ -87,7 +90,7 @@ class App: OdoodProgram {
         this.summary("Easily manage odoo installations.");
         this.topicGroup("Main");
         this.add(new CommandInit());
-        this.add(new CommandDeploy());
+        version(linux) this.add(new CommandDeploy());
         this.add(new CommandServer());
         this.add(new CommandStatus());
         this.add(new CommandDatabase());