+++ subpackages/lib/source/odood/lib/deploy/config.d
@@ -17,7 +17,8 @@ private import odood.lib.project.config:
     ProjectConfigDirectories,
     ProjectConfigOdoo;
 private import odood.lib.deploy.exception: OdoodDeployException;
-private import odood.lib.deploy.utils: dpkgCheckPackageInstalled;
+private import odood.lib.deploy.utils: checkSystemUserExists;
+private import odood.lib.venv: VenvOptions;
 private import odood.utils.odoo.serie: OdooSerie;
 private import odood.utils: generateRandomString;
 
@@ -52,8 +53,7 @@ struct DeployConfigOdoo {
 
 struct DeployConfig {
     Path deploy_path = Path("/", "opt", "odoo");
-    string py_version="auto";
-    string node_version="lts";
+    VenvOptions venv_options;
     OdooInstallType install_type=OdooInstallType.Archive;
 
     DeployConfigDatabase database;
@@ -106,9 +106,10 @@ struct DeployConfig {
             "Password for database must not be empty!");
 
         if (this.database.local_postgres)
+            // If local postgres requested, we expect that `postgres` user exists in system.
             enforce!OdoodDeployException(
-                !dpkgCheckPackageInstalled("potgresql"),
-                "Local postgres requested, but 'postgresql' package is not installed!");
+                checkSystemUserExists("postgres"),
+                "Local postgres requested, but 'postgresql' package seems not installed!");
     }
 
     /** Prepare odoo configuration file for this deployment