+++ subpackages/lib/source/odood/lib/venv.d
@@ -7,16 +7,18 @@ private import std.exception: enforce;
 private import std.conv: to;
 private import std.parallelism: totalCPUs;
 private import std.regex: ctRegex, matchFirst;
+private import std.string: strip;
 
 private static import std.process;
 
+private import theprocess;
 private import thepath: Path;
 private static import dyaml;
 
 private import odood.exception: OdoodException;
-private import theprocess;
 private import odood.utils;
 private import odood.utils.versioned: Version;
+private import odood.lib.odoo.python: getSystemPythonVersion;
 
 // TOOD: May be it have sense to move this to utils subpackage.
 
@@ -35,6 +37,42 @@ enum PySerie {
     py2=2,
     py3=3,
 }
+
+
+/// Python installation type
+enum PyInstallType {
+    System,
+    Build,
+    PyEnv,
+}
+
+
+/// Python installation options
+struct VenvOptions {
+    // TODO: Convert parameters to properties, to implement logic, that will
+    //       automatically change install type to build if version is set
+
+    // By default we use system python
+    PyInstallType install_type=PyInstallType.System;
+
+    // No version specified, if system python is in use
+    string py_version="";
+
+    // Default node version to install
+    string node_version="lts";
+}
+
+
+/// Return interpreter name for specified python serie
+@safe const(string) getPyInterpreterName(in PySerie py_serie) {
+    final switch(py_serie) {
+        case PySerie.py2:
+            return "python2";
+        case PySerie.py3:
+            return "python3";
+    }
+}
+
 
 /* TODO: move to utils package?
  *       Dyaml integration should be kept in lib
@@ -77,12 +115,7 @@ const struct VirtualEnv {
 
     /// Name of python interpreter
     @safe const(string) py_interpreter_name() const {
-        final switch(_py_serie) {
-            case PySerie.py2:
-                return "python2";
-            case PySerie.py3:
-                return "python3";
-        }
+        return getPyInterpreterName(_py_serie);
     }
 
     /// Python version for this venv
@@ -250,7 +283,8 @@ const struct VirtualEnv {
 
     /// ditto
     void buildPython(in Version build_version,
-                     in bool enable_sqlite=false) {
+                     in bool enable_sqlite=false,
+                     in bool enable_optimizations=false) {
 
         infof("Building python version %s...", build_version);
 
@@ -269,8 +303,10 @@ const struct VirtualEnv {
                     build_version, build_version);
         auto python_download_path = getCacheDir("python", tmp_dir).join(
             "python-%s.tgz".format(build_version));
-        auto python_build_dir = tmp_dir.join(
+        auto python_src_dir = tmp_dir.join(
             "Python-%s".format(build_version));
+        auto python_build_dir = tmp_dir.join(
+            "build");
         auto python_path = _path.join("python");
 
         // Ensure we can start building process
@@ -278,13 +314,19 @@ const struct VirtualEnv {
             !python_path.exists,
             "Python destination path (%s) already exists".format(python_path));
         enforce!OdoodException(
-            !python_build_dir.exists,
-            "Python build dir (%s) already exists.".format(python_build_dir));
+            !python_src_dir.exists,
+            "Python sources dir (%s) already exists.".format(python_src_dir));
 
-        string[] python_configure_opts = ["--prefix=%s".format(python_path)];
+        string[] python_configure_opts = [
+            "--prefix=%s".format(python_path),
+            "--srcdir=%s".format(python_src_dir),
+        ];
 
         if (enable_sqlite)
             python_configure_opts ~= "--enable-loadable-sqlite-extensions";
+
+        if (enable_optimizations)
+            python_configure_opts ~= "--enable-optimizations";
 
         if (!python_download_path.exists) {
             infof(
@@ -294,7 +336,7 @@ const struct VirtualEnv {
             download(python_download_link, python_download_path);
         }
 
-        if (!python_build_dir.exists) {
+        if (!python_src_dir.exists) {
             // Extract only if needed
             info("Unpacking python...");
             Process("tar")
@@ -304,9 +346,12 @@ const struct VirtualEnv {
                 .ensureStatus(true);
         }
 
+        // Ensure build dir exists
+        python_build_dir.mkdir(true);
+
         // Configure python build
         info("Running python's configure script...");
-        Process(python_build_dir.join("configure"))
+        Process(python_src_dir.join("configure"))
             .withArgs(python_configure_opts)
             .inWorkDir(python_build_dir)
             .execute()
@@ -389,60 +434,103 @@ const struct VirtualEnv {
       *         specific python version to build.
       *     node_version = NodeJS version to install.
       **/
-    void initializeVirtualEnv(
-            in string python_version,
-            in string node_version) {
+    void initializeVirtualEnv(in VenvOptions opts) {
         info("Installing virtualenv...");
 
-        if (python_version == "system") {
-            final switch(_py_serie) {
-                case PySerie.py2:
-                    Process("python3")
-                        .withArgs([
-                            "-m", "virtualenv",
-                            "-p", "python2",
-                            _path.toString])
-                        .execute()
-                        .ensureStatus(true);
-                    break;
-                case PySerie.py3:
-                    Process("python3")
-                        .withArgs([
-                            "-m", "virtualenv",
-                            "-p", "python3",
-                            _path.toString])
-                        .execute()
-                        .ensureStatus(true);
-                    break;
-            }
-        } else {
-            buildPython(python_version);
+        final switch(opts.install_type) {
+            case PyInstallType.System:
+                infof("Using system python (version: %s)", _py_serie.getSystemPythonVersion);
+                final switch(_py_serie) {
+                    case PySerie.py2:
+                        Process("python3")
+                            .withArgs([
+                                "-m", "virtualenv",
+                                "-p", "python2",
+                                _path.toString])
+                            .execute()
+                            .ensureStatus(true);
+                        break;
+                    case PySerie.py3:
+                        Process("python3")
+                            .withArgs([
+                                "-m", "virtualenv",
+                                "-p", "python3",
+                                _path.toString])
+                            .execute()
+                            .ensureStatus(true);
+                        break;
+                }
+                break;
+            case PyInstallType.Build:
+                buildPython(opts.py_version);
 
-            // Install virtualenv inside built python environment
-            Process(_path.join("python", "bin", "pip"))
-                .withArgs("install", "virtualenv")
-                .execute()
-                .ensureStatus(true);
+                // Install virtualenv inside built python environment
+                Process(_path.join("python", "bin", "pip"))
+                    .withArgs("install", "virtualenv")
+                    .execute()
+                    .ensureStatus(true);
 
-            // Initialize virtualenv inside built python env
-            Process(_path.join("python", "bin", "python"))
-                .withArgs([
-                    "-m", "virtualenv",
-                    "-p", _path.join("python", "bin", "python").toString,
-                    _path.toString])
-                .execute()
-                .ensureStatus(true);
+                // Initialize virtualenv inside built python env
+                Process(_path.join("python", "bin", "python"))
+                    .withArgs([
+                        "-m", "virtualenv",
+                        "-p", _path.join("python", "bin", "python").toString,
+                        _path.toString])
+                    .execute()
+                    .ensureStatus(true);
+                break;
+            case PyInstallType.PyEnv:
+                auto pyenv = resolveProgram("pyenv");
+                enforce!OdoodException(!pyenv.isNull, "pyenv not available!");
+                auto pyenv_path = pyenv.get();
+
+                infof("Using python %s via pyenv...", opts.py_version);
+
+                // Install desired python version (if needed)
+                Process(pyenv_path)
+                    .withArgs("install", "--skip-existing", opts.py_version)
+                    .execute
+                    .ensureOk(true);
+
+                // Find the prefix of installed (or existing) python of desired version
+                Path python_prefix = Process(pyenv_path)
+                    .withArgs("prefix", opts.py_version)
+                    .execute
+                    .ensureOk(true)
+                    .output
+                    .strip;
+
+                infof("Using python %s available at prefix %s...", opts.py_version, python_prefix.toString);
+
+                // Ensure virtualenv is installed in this pyenv python version
+                Process(python_prefix.join("bin", "pip"))
+                    .withArgs("install", "virtualenv")
+                    .execute
+                    .ensureOk(true);
+
+                // Init virtualenv
+                infof(
+                    "Initializing virtualenv with python %s...",
+                    python_prefix.join("bin", "python").toString);
+                Process(python_prefix.join("bin", "python"))
+                    .withArgs([
+                        "-m", "virtualenv",
+                        "-p", python_prefix.join("bin", "python").toString,
+                        _path.toString])
+                    .execute()
+                    .ensureOk(true);
+                break;
         }
 
-        //// Add bash script to run any command in virtual env
+        // Add bash script to run any command in virtual env
         initRunInVenvScript();
 
         // Install nodeenv and node
-        infof("Installing nodejs version %s", node_version);
+        infof("Installing nodejs version %s", opts.node_version);
         installPyPackages("nodeenv");
         runE([
             "nodeenv", "--python-virtualenv", "--clean-src",
-            "--jobs", totalCPUs.to!string, "--node", node_version,
+            "--jobs", totalCPUs.to!string, "--node", opts.node_version,
         ]);
 
         info("VirtualEnv initialized successfully!");