+++ tests/basic.d
@@ -17,24 +17,43 @@ import odood.lib.odoo.config: OdooConfigBuilder;
 
 import odood.cli.utils: printLogRecord;
 
+import peque;
+
+
+/// Prepare virtualenv options for test
+auto getVenvOptions(in OdooSerie serie) {
+    import odood.lib.venv: PyInstallType;
+    import odood.lib.odoo.python: guessVenvOptions;
+
+    auto venv_options = serie.guessVenvOptions;
+
+    if (venv_options.install_type == PyInstallType.System)
+        return venv_options;
+
+    if (environment.get("ODOOD_PREFER_PY_INSTALL") == "build")
+        venv_options.install_type = PyInstallType.Build;
+    else if (environment.get("ODOOD_PREFER_PY_INSTALL") == "pyenv")
+        venv_options.install_type = PyInstallType.PyEnv;
+
+    return venv_options;
+}
 
 /// Create new database user in postgres db
 void createDbUser(in string user, in string password) {
-    import dpq.connection: Connection;
 
-    auto connection_str = "host=%s port=%s dbname=postgres user=%s password=%s".format(
-        environment.get("POSTGRES_HOST", "localhost"),
-        environment.get("POSTGRES_PORT", "5432"),
-        environment.get("POSTGRES_USER", "odoo"),
-        environment.get("POSTGRES_PASSWORD", "odoo"));
-
-    infof("Connecting to postgres via: %s", connection_str);
-    auto connection = Connection(connection_str);
-    scope(exit) connection.close();
+    string[string] connection_params = [
+        "host": environment.get("POSTGRES_HOST", "localhost"),
+        "port": environment.get("POSTGRES_PORT", "5432"),
+        "user": environment.get("POSTGRES_USER", "odoo"),
+        "password": environment.get("POSTGRES_PASSWORD", "odoo"),
+        "dbname": "postgres",
+    ];
+    infof("Connecting to postgres via: %s", connection_params);
+    auto connection = Connection(connection_params);
 
     auto res_user_exists = connection.exec(
         "SELECT EXISTS(SELECT 1 FROM pg_roles WHERE rolname='%s')".format(user));
-    if (res_user_exists[0][0].as!string != "t") {
+    if (res_user_exists[0][0].get!string != "t") {
         infof("Creating db user '%s' with password '%s' for tests", user, password);
         connection.exec(
             "CREATE USER \"%s\" WITH CREATEDB PASSWORD '%s';".format(
@@ -159,12 +178,12 @@ void testAddonsManagementBasic(in Project project) {
         .run();
     test_result.success.shouldBeTrue();
 
-    // Try to fetch bureaucrate-knowledge from odoo apps
-    project.addons.downloadFromOdooApps("bureaucrat_knowledge");
-    project.directories.addons.join("bureaucrat_knowledge").exists.shouldBeTrue;
-    project.directories.addons.join("bureaucrat_knowledge").isSymlink.shouldBeTrue;
-    project.directories.addons.join("bureaucrat_knowledge").readLink.shouldEqual(
-        project.directories.downloads.join("bureaucrat_knowledge"));
+    // Try to fetch web_responsive from odoo apps
+    project.addons.downloadFromOdooApps("web_responsive");
+    project.directories.addons.join("web_responsive").exists.shouldBeTrue;
+    project.directories.addons.join("web_responsive").isSymlink.shouldBeTrue;
+    project.directories.addons.join("web_responsive").readLink.shouldEqual(
+        project.directories.downloads.join("web_responsive"));
 
     // Test parsing addons-list.txt file
     auto parsed_addons = project.addons.parseAddonsList(
@@ -196,10 +215,10 @@ void testRunningScripts(in Project project) {
     with (project.databases.get(dbname)) {
         runSQLQuery(
             "SELECT name FROM res_partner WHERE id = 1"
-        ).get(0, 0).as!string.get.shouldEqual("Test SQL 72");
+        )[0][0].get!string.shouldEqual("Test SQL 72");
         runSQLQuery(
             "SELECT name FROM res_partner WHERE id = 2"
-        ).get(0, 0).as!string.get.shouldEqual("Test SQL 75");
+        )[0][0].get!string.shouldEqual("Test SQL 75");
     }
 
     // Run Python Script
@@ -210,10 +229,10 @@ void testRunningScripts(in Project project) {
     with (project.databases.get(dbname)) {
         runSQLQuery(
             "SELECT name FROM res_partner WHERE id = 1"
-        ).get(0, 0).as!string.get.shouldEqual("Test PY 41");
+        )[0][0].get!string.shouldEqual("Test PY 41");
         runSQLQuery(
             "SELECT name FROM res_partner WHERE id = 2"
-        ).get(0, 0).as!string.get.shouldEqual("Test PY 42");
+        )[0][0].get!string.shouldEqual("Test PY 42");
     }
 
     infof("Testing running scripts for %s. Complete: Ok.", project);
@@ -264,7 +283,7 @@ unittest {
             "odoo")
         .setHttp("localhost", "18069")
         .result();
-    project.initialize(odoo_conf);
+    project.initialize(odoo_conf, project.odoo.serie.getVenvOptions);
     project.save();
 
     // Test created project
@@ -276,9 +295,9 @@ unittest {
     //project.runBasicTests;
 
     /*
-     * TODO: Currently, because some addons used in tests are not ported to 17,
+     * TODO: Currently, because some addons used in tests are not ported to 18,
      *       we do not test addons management. But later, when that addons ported
-     *       we have to chage this and run tests for addons management for Odoo 17
+     *       we have to chage this and run tests for addons management for Odoo 18
      */
 
     // Test server management
@@ -314,7 +333,7 @@ unittest {
             "odoo")
         .setHttp("localhost", "17069")
         .result();
-    project.initialize(odoo_conf);
+    project.initialize(odoo_conf, project.odoo.serie.getVenvOptions);
     project.save();
 
     // Test created project
@@ -323,25 +342,7 @@ unittest {
     project.config_path.shouldEqual(temp_path.join("odood.yml"));
 
     // Run basic tests
-    //project.runBasicTests;
-
-    /*
-     * TODO: Currently, because some addons used in tests are not ported to 17,
-     *       we do not test addons management. But later, when that addons ported
-     *       we have to chage this and run tests for addons management for Odoo 17
-     */
-
-    // Test server management
-    testServerManagement(project);
-
-    // Test LOdoo Database operations
-    testDatabaseManagement(project);
-
-    // Test basic addons management
-    //testAddonsManagementBasic(project);
-
-    // Test running scripts
-    testRunningScripts(project);
+    project.runBasicTests;
 }
 
 
@@ -365,7 +366,7 @@ unittest {
             "odoo")
         .setHttp("localhost", "16069")
         .result();
-    project.initialize(odoo_conf);
+    project.initialize(odoo_conf, project.odoo.serie.getVenvOptions);
     project.save();
 
     // Test created project
@@ -378,6 +379,7 @@ unittest {
 }
 
 
+version(linux)
 @("Basic Test Odoo 15")
 unittest {
     auto temp_path = createTempPath(
@@ -398,7 +400,7 @@ unittest {
             "odoo")
         .setHttp("localhost", "15069")
         .result();
-    project.initialize(odoo_conf);
+    project.initialize(odoo_conf, project.odoo.serie.getVenvOptions);
     project.save();
 
     // Test created project
@@ -411,6 +413,7 @@ unittest {
 }
 
 
+version(linux)
 @("Basic Test Odoo 14")
 unittest {
     auto temp_path = createTempPath(
@@ -431,7 +434,7 @@ unittest {
             "odoo")
         .setHttp("localhost", "14069")
         .result();
-    project.initialize(odoo_conf);
+    project.initialize(odoo_conf, project.odoo.serie.getVenvOptions);
     project.save();
 
     // Test created project
@@ -444,6 +447,7 @@ unittest {
 }
 
 
+version(linux)
 @("Basic Test Odoo 13")
 unittest {
     auto temp_path = createTempPath(
@@ -464,7 +468,7 @@ unittest {
             "odoo")
         .setHttp("localhost", "13069")
         .result();
-    project.initialize(odoo_conf);
+    project.initialize(odoo_conf, project.odoo.serie.getVenvOptions);
     project.save();
 
     // Test created project
@@ -477,6 +481,7 @@ unittest {
 }
 
 
+version(linux)
 @("Basic Test Odoo 12")
 unittest {
     auto temp_path = createTempPath(
@@ -497,7 +502,7 @@ unittest {
             "odoo")
         .setHttp("localhost", "12069")
         .result();
-    project.initialize(odoo_conf);
+    project.initialize(odoo_conf, project.odoo.serie.getVenvOptions);
     project.save();
 
     // Test created project
@@ -510,32 +515,32 @@ unittest {
 }
 
 
-@("Test resintall Odoo 14 to 15")
+@("Test resintall Odoo 16 to 17")
 unittest {
     auto temp_path = createTempPath(
         environment.get("TEST_ODOO_TEMP", std.file.tempDir),
-        "tmp-odood-14-15",
+        "tmp-odood-16-17",
     );
     scope(exit) temp_path.remove();
 
-    // Create database use for odoo 14 instance
-    createDbUser("odood14to15test", "odoo");
+    // Create database use for odoo 16 instance
+    createDbUser("odood16to17test", "odoo");
 
-    auto project = new Project(temp_path, OdooSerie(14));
+    auto project = new Project(temp_path, OdooSerie(16));
     auto odoo_conf = OdooConfigBuilder(project)
         .setDBConfig(
             environment.get("POSTGRES_HOST", "localhost"),
             environment.get("POSTGRES_PORT", "5432"),
-            "odood14to15test",
+            "odood16to17test",
             "odoo")
-        .setHttp("localhost", "14169")
+        .setHttp("localhost", "16169")
         .result();
-    project.initialize(odoo_conf);
+    project.initialize(odoo_conf, project.odoo.serie.getVenvOptions);
     project.save();
 
     // Test created project
     project.project_root.shouldEqual(temp_path);
-    project.odoo.serie.shouldEqual(OdooSerie(14));
+    project.odoo.serie.shouldEqual(OdooSerie(16));
     project.config_path.shouldEqual(temp_path.join("odood.yml"));
 
 
@@ -545,12 +550,12 @@ unittest {
     // Test that server initialization works fine
     project.server.getServerRunner("--stop-after-init", "--no-http").execute;
 
-    // Reinstall Odoo to version 15
-    project.reinstallOdoo(OdooSerie(15), true);
+    // Reinstall Odoo to version 17
+    project.reinstallOdoo(OdooSerie(17), true);
 
-    // Test project after Odoo reinstalled to version 15
+    // Test project after Odoo reinstalled to version 17
     project.project_root.shouldEqual(temp_path);
-    project.odoo.serie.shouldEqual(OdooSerie(15));
+    project.odoo.serie.shouldEqual(OdooSerie(17));
     project.config_path.shouldEqual(temp_path.join("odood.yml"));
 
     // Test that server initialization works fine